+++ astroquery/alma/tests/test_alma.py
@@ -173,6 +173,16 @@ def test_gen_str_sql():
         "(proposal_id LIKE '2012.%' OR proposal_id LIKE '2013._3%')"
 
 
+def test_gen_array_sql():
+    # test string array input (regression in #2094)
+    # string arrays should be OR'd together
+    common_select = "select * from ivoa.obscore WHERE "
+    test_keywords = ["High-mass star formation", "Disks around high-mass stars"]
+    assert _gen_sql({"spatial_resolution": "<0.1",
+        "science_keyword": test_keywords}) == common_select + \
+            "spatial_resolution<=0.1 AND (science_keyword='High-mass star formation' OR science_keyword='Disks around high-mass stars')"
+
+
 def test_gen_datetime_sql():
     common_select = 'select * from ivoa.obscore WHERE '
     assert _gen_sql({'start_date': '01-01-2020'}) == common_select + \